rewrite mtconnect2mqtt module
start from class MTConnect2MQTT
parameterize source agent and target broker
streamline storage of nextSequence and different cases for gathering response documnets
