+++ .idea/modules.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/mtconnectToMQTT/.idea/DataPipelineTestArchitecture.iml" filepath="$PROJECT_DIR$/mtconnectToMQTT/.idea/DataPipelineTestArchitecture.iml" />
+      <module fileurl="file://$PROJECT_DIR$/mtconnectToMQTT/mtconnectToMQTT.iml" filepath="$PROJECT_DIR$/mtconnectToMQTT/mtconnectToMQTT.iml" />
+    </modules>
+  </component>
+</project>
