+++ mtconnect2mqtt/.gitignore
@@ -1,0 +1,3 @@
+# Default ignored files
+
+
