+++ mtconnect2mqtt/.idea/.name
@@ -1,0 +1,1 @@
+mtconnect2MQTT
