+++ mtconnect2mqtt/.idea/libraries/junit_junit_4_12.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="junit:junit:4.12" type="repository">
+    <properties maven-id="junit:junit:4.12" />
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.m2/junit-4.12.jar!/" />
+      <root url="jar://$USER_HOME$/.m2/hamcrest-core-1.3.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
