+++ mtconnect2mqtt/.idea/libraries/org_eclipse_paho_org_eclipse_paho_client_mqttv3_1_2_2.xml
@@ -1,0 +1,12 @@
+<component name="libraryTable">
+  <library name="org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.2.2" type="repository">
+    <properties maven-id="org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.2.2" />
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.m2/org.eclipse.paho.client.mqttv3-1.2.2.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.m2/org.eclipse.paho.client.mqttv3-1.2.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
