+++ mtconnect2mqtt/.idea/libraries/org_hamcrest_hamcrest_all_1_3.xml
@@ -1,0 +1,10 @@
+<component name="libraryTable">
+  <library name="org.hamcrest:hamcrest-all:1.3" type="repository">
+    <properties maven-id="org.hamcrest:hamcrest-all:1.3" />
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.m2/hamcrest-all-1.3.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
