+++ mtconnect2mqtt/mtconnect2mqtt.iml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" inherit-compiler-output="true">
+    <exclude-output />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
+      <excludeFolder url="file://$MODULE_DIR$/src/deprecated" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="org.eclipse.paho.client.mqttv3-1.2.2-sources" level="project" />
+    <orderEntry type="library" name="org.hamcrest:hamcrest-all:1.3" level="project" />
+    <orderEntry type="library" name="junit:junit:4.12" level="project" />
+    <orderEntry type="library" name="org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.2.2" level="project" />
+  </component>
+</module>
