+++ mtconnectsourceconnector/.idea/libraries/kafka_connect_mqtt_1_2_3.xml
@@ -1,0 +1,9 @@
+<component name="libraryTable">
+  <library name="kafka-connect-mqtt-1.2.3">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/Desktop/kafka-connect-mqtt-1.2.3.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
