--- src/main/test-doc.xml
@@ -1,6 +1,0 @@
-<Header>
-    <Name>Mitch Hedburg</Name>
-    <Occupation>Comedian</Occupation>
-    <Address>123 Fake St.</Address>
-    <Slogan>If you're flammable and have legs, you are never blocking a fire exit.</Slogan>
-</Header>
